+++ imod/mf6/__init__.py
@@ -13,6 +13,7 @@ from imod.mf6.dsp import Dispersion
 from imod.mf6.evt import Evapotranspiration
 from imod.mf6.ghb import GeneralHeadBoundary
 from imod.mf6.hfb import (
+    HorizontalFlowBarrierBase,
     HorizontalFlowBarrierHydraulicCharacteristic,
     HorizontalFlowBarrierMultiplier,
     HorizontalFlowBarrierResistance,
@@ -32,6 +33,7 @@ from imod.mf6.npf import NodePropertyFlow
 from imod.mf6.oc import OutputControl
 from imod.mf6.out import open_cbc, open_hds, read_cbc_headers, read_grb
 from imod.mf6.rch import Recharge
+from imod.mf6.regridding_utils import RegridderInstancesCollection, RegridderType
 from imod.mf6.riv import River
 from imod.mf6.simulation import Modflow6Simulation
 from imod.mf6.src import MassSourceLoading
@@ -39,5 +41,5 @@ from imod.mf6.ssm import SourceSinkMixing
 from imod.mf6.sto import SpecificStorage, Storage, StorageCoefficient
 from imod.mf6.timedis import TimeDiscretization
 from imod.mf6.uzf import UnsaturatedZoneFlow
-from imod.mf6.wel import WellDisStructured, WellDisVertices
+from imod.mf6.wel import Well, WellDisStructured, WellDisVertices
 from imod.mf6.write_context import WriteContext